Update coding tests
